host/src/pyproject.toml
[project]
name = "book-code-agents"
version = "0.1.0"
description = "AI agents using OpenAI and MCP"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.9.0",
    "openai>=1.79.0",
    "python-dotenv>=1.0.0",
]
